Update state-map.txt
I found this name mapping manually. The name mapping does not appear to be in the Yosys log.
I'll look into this further, if we change the verbosity level, it might print this out during
the pmuxtree pass. Alternatively, we could read in the unflattened fabric first, just to identify
the signal name how I did manually.